requirements.txt
fastapi
langchain
openai
mysql-connector-python
sqlalchemy
twilio
dotenv